+++ src/core/GeoTech.ts
@@ -1,22 +1,32 @@
 /* qeslint-disable*/
-import { HeadingPitchRange, Event } from "cesium";
+import { BoundingSphere, Cartesian3, HeadingPitchRange, Event, Math as CesiumMath } from "cesium";
 import { GeoTechViewer } from "./GeoTechViewer";
 import { Site } from "./common";
 import { CanvasEventHandler } from "./CanvasEventHandler";
 import { MapTool } from "./MapTool";
+import { Area } from "./Area";
+import { AreaManager } from "./AreaManager";
 import APIInterface from "./APIInterface";
 import UIManager from "./UIManager";
 import NavigationHelper from "./NavigationHelper";
 
 export class GeoTech {
     readonly rootElementId = "root";
+
     readonly mainViewer = new GeoTechViewer(this);
+
     readonly apiInterface = new APIInterface();
+
     private _mapTool: MapTool | undefined;
+
+    private _areaManager: AreaManager | undefined;
+
     private _canvasEventHandler: CanvasEventHandler | undefined;
 
     protected readonly _mapToolDeactivated: Event = new Event();
+
     public uiManager = new UIManager();
+
     private _navigationHelper: NavigationHelper | undefined;
 
     constructor() {
@@ -33,6 +43,10 @@ export class GeoTech {
 
     get mapViewer() {
         return this.mainViewer.geoTechMapViewer!;
+    }
+
+    get areaManager() {
+        return this._areaManager as AreaManager;
     }
 
     get scene() {
@@ -101,6 +115,11 @@ export class GeoTech {
         this._navigationHelper = new NavigationHelper({
             viewer: viewer
         });
+
+        this._areaManager = new AreaManager({
+            geoTech: this,
+            scene: this.scene
+        });
     }
 
     zoomToSite(site: Site) {
@@ -110,6 +129,23 @@ export class GeoTech {
             if (entity.properties && entity.properties.name && entity.properties.name.getValue() === site.entity_name) {
                 viewer.zoomTo(entity, new HeadingPitchRange(viewer.scene.camera.heading, -0.1, 1000));
             }
+        });
+    }
+
+    zoomToArea(area: Area) {
+        const viewer = this.viewer;
+        area.showHide(true);
+
+        // Set the default pitch to 0 degrees
+        const pitch = CesiumMath.toRadians(0);
+
+        // Calculate the range to achieve an elevation of 109
+        const targetElevation = 50; // Desired elevation
+        const poiPosition = area.getPosition(new Cartesian3());
+        const range = targetElevation; // Adjust range based on POI's elevation
+
+        viewer.camera.flyToBoundingSphere(new BoundingSphere(poiPosition, 2), {
+            offset: new HeadingPitchRange(0, pitch, range) // Heading: 0, Pitch: 0, Range: calculated
         });
     }
 }