+++ src/core/GeoTechMapViewer.ts
@@ -1,34 +1,51 @@
 // qts-nocheck
 /* qeslint-disable */
 import {
+    Billboard,
+    BlendOption,
     Camera,
+    Cartesian2,
     Cartesian3,
-    Cesium3DTileset,
     defined,
     DeveloperError,
+    HeadingPitchRoll,
+    HorizontalOrigin,
     Ion,
     Math as CesiumMath,
     Matrix4,
-    HeadingPitchRoll,
+    Label,
+    LabelCollection,
+    LabelStyle,
     Rectangle,
-    Viewer,
     // @ts-ignore
     subscribeAndEvaluate,
     // @ts-ignore
     createDefaultImageryProviderViewModels,
     ProviderViewModel,
-    Primitive
+    Primitive,
+    ScreenSpaceEventHandler,
+    ScreenSpaceEventType,
+    VerticalOrigin,
+    Viewer,
+    Color
 } from "cesium";
 import viewerCesiumNavigationMixin from "./cesium-navigation-es6/viewerCesiumNavigationMixin";
 import { GeoTech } from "./GeoTech";
+import { defaultHeight } from "./Area";
 import { CapturedCameraProps } from "./common";
 
 export class GeoTechMapViewer {
     private readonly _geoTech: GeoTech;
+
     private readonly _viewer: Viewer;
+
     private _customTimePrimitive: Primitive | null = null;
+
     private _currentCameraProps: CapturedCameraProps | null = null;
+
     private _baseImageryProviders: any;
+
+    private _hoverLabel: Label;
 
     constructor(container: HTMLElement, geoTech: GeoTech) {
         // >>includeStart('debug', pragmas.debug);
@@ -45,7 +62,7 @@ export class GeoTechMapViewer {
         this._geoTech = geoTech;
 
         this._baseImageryProviders = createDefaultImageryProviderViewModels();
-        const selectedImageryProviderIndex = 1;
+        const selectedImageryProviderIndex = 0;
 
         this._viewer = new Viewer(container, {
             timeline: false,
@@ -108,13 +125,15 @@ export class GeoTechMapViewer {
         /**
          * we can hide the globe but if we do so, user can not show blue sky
          */
-        viewer.scene.globe.show = true;
-        viewer.scene.skyAtmosphere.show = true;
-        viewer.scene.highDynamicRange = true;
+        const scene = viewer.scene;
+
+        scene.globe.show = true;
+        scene.skyAtmosphere.show = true;
+        scene.highDynamicRange = true;
 
         // make sure user shows blue sky
-        viewer.scene.globe.enableLighting = false;
-        viewer.scene.globe.atmosphereLightIntensity = 20.0;
+        scene.globe.enableLighting = false;
+        scene.globe.atmosphereLightIntensity = 20.0;
 
         const camera = viewer.camera;
         camera.percentageChanged = 0.01; // default
@@ -131,6 +150,12 @@ export class GeoTechMapViewer {
             }
         });
 
+        camera.moveStart.addEventListener(() => {
+            if (this._hoverLabel) {
+                this._hoverLabel.show = false;
+            }
+        });
+
         if (viewerCesiumNavigationMixin) {
             const options = {
                 enableCompass: true,
@@ -142,6 +167,62 @@ export class GeoTechMapViewer {
 
             viewerCesiumNavigationMixin(viewer, options);
         }
+
+        const labelCollection = new LabelCollection({
+            blendOption: BlendOption.OPAQUE_AND_TRANSLUCENT
+        });
+
+        scene.primitives.add(labelCollection);
+
+        this._hoverLabel = labelCollection.add({
+            show: false,
+            text: "",
+            horizontalOrigin: HorizontalOrigin.CENTER,
+            verticalOrigin: VerticalOrigin.BOTTOM,
+            position: Cartesian3.fromDegrees(0, 0, 0),
+            eyeOffset: new Cartesian3(0, 0, 0),
+            scale: 0.8,
+            fillColor: Color.YELLOW,
+            outlineColor: Color.BLACK,
+            outlineWidth: 2,
+            style: LabelStyle.FILL_AND_OUTLINE,
+            disableDepthTestDistance: Number.POSITIVE_INFINITY
+        });
+
+        const handler = new ScreenSpaceEventHandler(scene.canvas);
+
+        handler.setInputAction((movement: { endPosition: Cartesian2 }) => {
+            const pickedObject = scene.pick(movement.endPosition);
+
+            if (!pickedObject) {
+                this._hoverLabel.show = false;
+                return;
+            }
+
+            if (!pickedObject.primitive) {
+                this._hoverLabel.show = false;
+                return;
+            }
+
+            const primitive = pickedObject.primitive;
+
+            if (!(primitive instanceof Billboard)) {
+                this._hoverLabel.show = false;
+                return;
+            }
+
+            const id = primitive.id;
+
+            const area = this._geoTech.areaManager.getAreaByCityName(id);
+
+            const longitude = area.longitude;
+            const latitude = area.latitude;
+            const height = defaultHeight;
+
+            this._hoverLabel.position = Cartesian3.fromDegrees(longitude, latitude, height);
+            this._hoverLabel.text = `${area.name} : ${area.geWeatherStatus()}`;
+            this._hoverLabel.show = true;
+        }, ScreenSpaceEventType.MOUSE_MOVE);
     }
 
     get viewer() {