+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### 1.0.3 - 2025-02-24
+
+#### Added
+
+-   Added area weather model and label for each area
+-   Added area manager class for area management
+
 ### 1.0.2 - 2025-02-21
 
 #### Updated